WIP: Create Repository class instead of JsonLoad
