--- app/src/main/java/com/github/amusinamaria/repository/data/JsonLoad.kt
@@ -1,83 +1,0 @@
-package com.github.amusinamaria.repository.data
-
-import android.content.Context
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.coroutineScope
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-import kotlinx.serialization.decodeFromString
-import kotlinx.serialization.json.Json
-
-private val jsonFormat = Json { ignoreUnknownKeys = true }
-
-private suspend fun loadGenres(context: Context): List<Genre> = withContext(Dispatchers.IO) {
-    val data = readAssetFileToString(context, "genres.json")
-    parseGenres(data)
-}
-
-private fun parseGenres(data: String): List<Genre> {
-    return jsonFormat.decodeFromString(data)
-}
-
-private fun readAssetFileToString(context: Context, fileName: String): String {
-    val stream = context.assets.open(fileName)
-    return stream.bufferedReader().readText()
-}
-
-private suspend fun loadActors(context: Context): List<Actor> = withContext(Dispatchers.IO) {
-    val data = readAssetFileToString(context, "people.json")
-    parseActors(data)
-}
-
-private fun parseActors(data: String): List<Actor> {
-    return jsonFormat.decodeFromString(data)
-}
-
-internal suspend fun loadMovies(context: Context): List<Movie> = withContext(Dispatchers.IO) {
-    var genresMap = listOf<Genre>()
-    var actorsMap = listOf<Actor>()
-    var data = ""
-    coroutineScope {
-        launch {
-            genresMap = loadGenres(context)
-        }
-        launch {
-            actorsMap = loadActors(context)
-        }
-        launch {
-            data = readAssetFileToString(context, "data.json")
-        }
-    }
-    parseMovies(data, genresMap, actorsMap)
-}
-
-private fun parseMovies(
-    data: String,
-    genres: List<Genre>,
-    actors: List<Actor>
-): List<Movie> {
-    val genresMap = genres.associateBy { it.id }
-    val actorsMap = actors.associateBy { it.id }
-
-    val jsonMovies = jsonFormat.decodeFromString<List<JsonMovie>>(data)
-
-    return jsonMovies.map { jsonMovie ->
-        Movie(
-            id = jsonMovie.id,
-            title = jsonMovie.title,
-            overview = jsonMovie.overview,
-            poster = jsonMovie.posterPicture,
-            backdrop = jsonMovie.backdropPicture,
-            ratings = jsonMovie.ratings,
-            numberOfRatings = jsonMovie.votesCount,
-            minimumAge = if (jsonMovie.adult) 16 else 13,
-            runtime = jsonMovie.runtime,
-            genres = jsonMovie.genreIds.map {
-                genresMap[it] ?: throw IllegalArgumentException("Genre not found")
-            },
-            actors = jsonMovie.actors.map {
-                actorsMap[it] ?: throw IllegalArgumentException("Actor not found")
-            }
-        )
-    }
-}
